// desktop/src/types.ts
type TMaybe<T> = T | null | undefined
// ...
export type TImportWorkspaceConfig = Readonly<{
  workspaceID: string
  workspaceUID: string
  devsyProHost: string
  project: string
  options: Record<string, string> | null
}>
// ...
export type TProID = string
export type TWithProID = Readonly<{ id: TProID }>
export type TProInstance = Readonly<{
  host: TMaybe<string>
  provider: TMaybe<string>
  creationTimestamp: TMaybe<string>
  authenticated: TMaybe<boolean>
  capabilities: TMaybe<readonly string[]>
}>
// ...
export type TListProInstancesConfig = Readonly<
  | {
      authenticate?: boolean
    }
  | undefined
>
export type TPlatformVersionInfo = Readonly<{
  serverVersion: TMaybe<string>
  remoteProviderVersion: TMaybe<string>
  currentProviderVersion: TMaybe<string>
}>
//#endregion

export type TDevcontainerSetup = Readonly<{
  isGitRepository: boolean
  isLocal: boolean
  isImage: boolean
  configPaths: string[]
}>
//#region CommunityContributions
export type TCommunityContributions = Readonly<{
  providers: readonly TCommunityProvider[]
}>
export type TCommunityProvider = Readonly<{
  repository: string
}>
//#endregion
export type TPlatformHealthCheck = Readonly<{
  healthy: TMaybe<boolean>
  online: TMaybe<boolean>
  loginRequired: TMaybe<boolean>
  details: TMaybe<string[]>
}>
export type TPlatformUpdateCheck = Readonly<{
  available: TMaybe<boolean>
  newVersion: TMaybe<string>
}>